Bugfix: BUGS-4 / RLAPI-G2 (Undefined variable: check_against_hashlist)

+++ src/Uploader/Uploader.php
@@ -177,11 +177,13 @@ class Uploader
                 ],
             ];
         }
-        if (true == $check_against_hashlist['clearance']) {
-            if (null != $file_name) {
-                unlink(getenv('TMP_STORE').$file_name);
-            } else {
-                unlink(implode('', $file['tmp_name']));
+        if($authentication){
+            if (true == $check_against_hashlist['clearance']) {
+                if (null != $file_name) {
+                    unlink(getenv('TMP_STORE').$file_name);
+                } else {
+                    unlink(implode('', $file['tmp_name']));
+                }
             }
         }
